Temporary fix for the event reference retrieval in parts of the code

+++ CepGen/Core/GridParameters.cpp
@@ -22,7 +22,7 @@ namespace cepgen
     for ( size_t i = 0; i < max_; ++i ) {
       size_t jj = i;
       for ( size_t j = 0; j < ndim; ++j ) {
-        unsigned int tmp = jj*INV_M_BIN;
+        size_t tmp = jj*INV_M_BIN;
         //coord[j] = roundf( jj-tmp*M_BIN );
         coord[j] = jj-tmp*M_BIN;
         jj = tmp;
